服务器接收消息与pb

+++ GoServer/ziface/imessage.go
@@ -18,10 +18,10 @@ package ziface
 */
 type IMessage interface {
 	GetDataLen() uint32 //获取消息数据段长度
-	GetMsgID() uint32   //获取消息ID
+	GetMsgID() uint16   //获取消息ID
 	GetData() []byte    //获取消息内容
 
-	SetMsgID(uint32)   //设计消息ID
+	SetMsgID(uint16)   //设计消息ID
 	SetData([]byte)    //设计消息内容
 	SetDataLen(uint32) //设置消息数据段长度
 }
